Fix failing tests and enable running tests on JVM
This commit enables a JVM-based workflow. The tests that pass without
modification on a JVM can now be run on a JVM. The tests that fail on JVM now
only run on Babashka.

In addition, test data has been updated to reflect the reality - per 2024-11-27,
the latest version of org.babashka/http-server is 0.1.13.