Also remove Launchpad
The way I'd added Launchpad in a previous commit in this PR would be bad for all
bbin users: they would all be forced to download Launchpad. I don't want that.
Launchpad is a development helper, it shouldn't carry weight that influences
bbin users badly.

So I'm removing Launchpad.

We could possibly get around shipping Launchpad to bbin users by loading
launchpad dynamically in the `bb launchpad` task, but that's not worth it for me
for now.